+++ src/main/java/com/moilioncircle/redis/rdb/cli/cmd/XRmt.java
@@ -61,8 +61,8 @@ import picocli.CommandLine.Spec;
 		optionListHeading = "%nOptions:%n",
 		versionProvider = XVersionProvider.class,
 		customSynopsis = {
-				"Usage: rmt [-hV] -s <source> (-m <uri> | -c <config>) [-d <db>...]",
-				"       [-k <regex>...] [-t <type>...] [-r] [-l]"
+				"Usage: rmt [-hV] -s <source> (-m <uri> | -c <conf>) [-d <db>...]",
+				"       [-k <regex>...] [-t <type>...] [-rl]"
 		},
 		footer = {"%nExamples:",
 				"  rmt -s ./dump.rdb -c ./nodes.conf -t string -r",
@@ -80,7 +80,7 @@ public class XRmt implements Callable<Integer> {
 		@Option(names = {"-m", "--migrate"}, required = true, paramLabel = "<uri>", description = {"Migrate to uri. eg:", "redis://host:port?authPassword=foobar."})
 		private String migrate;
 		
-		@Option(names = {"-c", "--config"}, required = true, description = {"Migrate data to cluster via redis cluster's", "<nodes.conf> file, if specified, no need to", "specify --migrate."}, type = File.class)
+		@Option(names = {"-c", "--config"}, required = true, paramLabel = "<conf>", description = {"Migrate data to cluster via redis cluster's", "<nodes.conf> file, if specified, no need to", "specify --migrate."}, type = File.class)
 		private File config;
 	}